# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ms gRPC Sample",
        "title": "Ms gRPC Sample",
        "contact": {
            "name": "Sefik Can Kanber",
            "url": "https://github.com/sefikcan"
        },
        "version": "1.0"
    },
    "host": "localhost:50050",
    "basePath": "/api/v1",
    "paths": {
        "/products": {
            "post": {
                "description": "Create product handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Create product",
                "parameters": [
                    {
                        "description": "Create Product",
                        "name": "createProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.ProductResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Get by id product handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get by id product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ProductResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update product handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Product",
                        "name": "updateProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ProductResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete by id product handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "requests.CreateProductRequest": {
            "type": "object",
            "required": [
                "category",
                "name",
                "optionName"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 3
                },
                "optionName": {
                    "type": "string"
                }
            }
        },
        "requests.UpdateProductRequest": {
            "type": "object",
            "required": [
                "category",
                "id",
                "name",
                "optionName"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 3
                },
                "optionName": {
                    "type": "string"
                }
            }
        },
        "responses.ProductResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optionName": {
                    "type": "string"
                }
            }
        }
    }
}